Extract TypedBlockData to typed_block_data.hpp, add Column accessors
Phase 3 step 1: Extract TypedBlockData, Enumeration, PointerWrapper,
IteratorType, IteratorDensity, and make_typed_block free function to a
new typed_block_data.hpp header. This eliminates duplication and prepares
for ColumnData removal.

Add sparse_map_ptr() and shapes_buffer() accessors to Column to expose
the data that ColumnData callers need without going through ColumnData.
